refactor/reflect PR feedback
- Simpler main Application
- Remove unnecessary local variable reassignments
- styling: one method per line in method chain
